Create new dataloader for raml
The raml_read data is in the dataloader. Created new get_raml_batch and get_next_raml_batch for this data. That way, the output of get_raml_batch is the same, and can be taken as input of get_batch of Seq2seq class (just adds the reward as atttribute of incoming).
Still need to be run to check.